REVERTME [VPG]: drm/i915/bdw: Move WaDisableFenceDestinationToSLM workarounds to render ring init
Some workarounds modify registers that are part of the register
state context, but these are not initialized until init_clock_gating().
This initialization happens after the default contexts have been
populated, so it is possible to loose the workarounds. This is also true
in case of a reset.
The fix is to move the workarounds from init_clock_gating() to the
render ring init.
Marked as REVERTME because this is only temporary. Workarounds should not
be in init_clock_gating() to begin with. This is being addressed and
will be sent upstream.
Issue: APDEV-3096
Change-Id: Ib8dde2deb4dcba87db5da00795906cad9ef8657d
Signed-off-by: Michel Thierry <michel.thierry@intel.com>